Use more direct STArray type in annotations instead of the more generic MArray
Both STArray and MArray refer to mutable arrays. The STArray type is
specific to the ST monad, while MArray is a type class that abstracts
over mutable arrays and can be used in various monads, including ST.

In our case, the more direct type signature is better for making the
code less overwhelming.

The 's' parameter is a phantom type representing the state thread, and
is used to ensure that the array is only modified in the context of
the ST monad.